_bodySize Type change int -> size_t

+++ include/Request/Request.hpp
@@ -12,7 +12,7 @@ class Request
 	std::map<std::string, std::string>	_env_cgi;
 	int									_ret;
 	std::string							_body;
-	int									_bodySize;
+	size_t								_bodySize;
 	std::string							_port;
 	std::string							_path;
 	std::string							_query;
